# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Kootenai cybersecurity education platform\nOrchestrates lab environments across Proxmox and CloudStack",
        "title": "Kootenai API",
        "contact": {
            "name": "Project repository",
            "url": "https://github.com/toddbartholow/kootenai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Get recent events with optional filtering by type, pod, session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of events (default 50, max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type (syscheck, audit, auth, ssh)",
                        "name": "eventType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by pod ID",
                        "name": "podId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by session ID",
                        "name": "sessionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by VM name",
                        "name": "vmName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 format)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 format)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.EventsResponse"
                        }
                    }
                }
            }
        },
        "/events/pod/{podID}": {
            "get": {
                "description": "Get all events for a specific pod",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get pod events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pod ID",
                        "name": "podID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of events (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.EventsResponse"
                        }
                    }
                }
            }
        },
        "/events/session/{sessionID}": {
            "get": {
                "description": "Get all events for a specific session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get session events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of events (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.EventsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server.EventSummary": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matchedCheckpoints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "podId": {
                    "type": "string"
                },
                "processed": {
                    "type": "boolean"
                },
                "ruleId": {
                    "type": "string"
                },
                "ruleLevel": {
                    "type": "integer"
                },
                "sessionId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "vmName": {
                    "type": "string"
                }
            }
        },
        "internal_server.EventsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server.EventSummary"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}